Revert "Changed from Float to Double"
This reverts commit ed0b47b01ca835c0a6f27a74b9d127e59725ec95.

+++ Back-End/Node.java
@@ -1,16 +1,16 @@
 public class Node {
     public int id;
     public String location;
-    public double latitude;
-    public double longitude;
+    public float latitude;
+    public float longitude;
     
     // Constructs a new node instance given a id, latitude, and longitude.
-    public Node(int id, double latitude, double longitude) {
+    public Node(int id, float latitude, float longitude) {
         this(id, latitude, longitude, null);
     }
 
     // Constructs a new node instance given a id, location, latitude, and longitude.
-    public Node(int id, double latitude, double longitude, String location) {
+    public Node(int id, float latitude, float longitude, String location) {
         this.id = id;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -28,12 +28,12 @@ public class Node {
     }
 
     // Returns a decimal latitude.
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
     // Returns a decimal longitude.
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
